// src/generated/exchange-booking.ts
export interface components {
    schemas: {
        ExchangeBookingRequest: {
            /** @description Contains the required parameters to make a booking exchange request. */
            ExchangeBookingRQ?: {
// ...
                AutomatedExchanges: {
// ...
                    ExchangeConfirmation?: {
// ...
                        OptionalQualifiers?: {
                            /** @description Contains all pricing qualifiers that may affect the final stored price quote reissue (PQR) record. */
                            FOP_Qualifiers?: {
                                /** @description Lists single form of payment information. Multiple forms of payment are not supported. */
                                BasicFOP: {
// ...
                                    CC_Info?: {
// ...
                                        PaymentCard?: {
                                            /**
                                             * @description The credit card vendor code.
                                             * @example VI
                                             */
                                            Code: string;
                                            /**
                                             * @description The credit card expiration date in YYYY-MM format.
                                             * @example 2012-12
                                             */
                                            ExpireDate: string;
                                            /**
                                             * @description The manual credit card approval code for credit cards that are accepted, but are not processed automatically.
                                             * @example 123456
                                             */
                                            ManualApprovalCode?: string;
                                            /**
                                             * @description The credit card number.
                                             * @example 4123412341234123
                                             */
                                            Number: string;
                                        };
                                    };
// ...
                                }[];
                            };
// ...
                        };
                    };
                }[];
            };
        };
        /** @description Contains all the elements of the response for Exchange Booking transaction. */
        ExchangeBookingResponse: {
// ...
            ExchangeBookingRS?: {
                ApplicationResults?: components["schemas"]["ApplicationResults"];
                /** @description Used to pass details pertaining to the user specified acceptable price difference. */
                ExchangeConfirmation?: {
                    /**
                     * @description Used to pass the PQR number.
                     * @example 02
                     */
                    PQR_Number?: string;
                    /**
                     * @description Used to return information if transaction was auto routed to fare category 16.
                     * @example Y
                     */
                    AutoRedirect?: string;
                    PriceComparison?: {
                        /**
                         * @description Used to return the total amount associated with pricing of the exchange. `-200.00` reflects the refund amount.
                         * @example -5.99
                         */
                        amountReturned?: string;
                    };
                    /**
                     * @description Used to pass information on the refundable amount on EMD.
                     * @example 199.00
                     */
                    RefundableAmountOnEMD?: string;
                    /**
                     * @description Used to return the additional free text information that applies to the confirmed PQR record.
                     * @example CHG FEE AMT MODIFIED - REISSUE GUARANTEE NO LONGER APPLIES
                     */
                    AdditionalText?: string;
                }[];
                /** @description Used to pass information on the generated PQR record. */
                PriceQuoteReissue?: {
// ...
                    MiscInformation?: {
                        BaggageFees?: {
                            /**
                             * @description Used to return baggage fee related information.
                             * @example BAG ALLOWANCE -LASLAX-NIL/DL
                             */
                            Text?: string;
                        }[];
                        SignatureLine?: {
                            /**
                             * @description Used to return the agent sine that stored the price.
                             * @example ASP
                             */
                            CreationAgent?: string;
                            /**
                             * @description Used to return creation date/time information.
                             * @example 2026-03-02T09:30
                             */
                            CreateDateTime?: string;
                            /**
                             * @description Used to return the base sine in location associated to the agent sine that stored the price.
                             * @example G7HE
                             */
                            HomePseudoCityCode?: string;
                            /**
                             * @description Used to return the location associated to the agent sine when they stored the price.
                             * @example G7HE
                             */
                            PseudoCityCode?: string;
                            /**
                             * @description Used to return the pricing source.
                             * @example SYS
                             */
                            Source?: string;
                        }[];
                    };
                    /** @description Used to collect priced itinerary information. */
                    PricedItinerary?: {
                        /**
                         * @description Used to return the pricing input command associated with the particular price quote record.
                         * @example WFRF
                         */
                        InputMessage?: string;
                        /** @description Used to return the document number. */
                        AirItineraryPricingInfo?: {
// ...
                            PassengerTypeQuantity?: {
                                /**
                                 * @description Used to return the passenger type code used to price the itinerary.
                                 * @example ADT
                                 */
                                Code?: string;
                                /**
                                 * @description Used to return the passenger name number.
                                 * @example 01
                                 */
                                Quantity?: string;
                            }[];
// ...
                        };
                    };
                }[];
            };
        };
// ...
        ProblemInformation: {
            /** @description Lists specific information from the system and/or other sub-systems invoked during processing. */
            SystemSpecificResults?: {
                HostCommand?: {
                    /**
                     * @description An abbreviation for Line / Interface Adapter / Terminal Address.
                     *
                     *     It is a 3-byte terminal address, usually seen as a 6-character hexadecimal field.
                     * @example 0358DC
                     */
                    LNIATA?: string;
                    /**
                     * @description Sabre host system command.
                     * @example 120MAYDFWLAX
                     */
                    content?: string;
                };
                Message?: {
                    /**
                     * @description An abbreviated version of the error or warning in textual format.
                     * @example WARN.SP.PROVIDER_WARNING
                     */
                    code?: string;
                    /**
                     * @description An indication of the source of error or warning when processing the request.
                     * @example No new price quote reissue (PQR) records have been stored, see remaining messages for details.
                     */
                    content?: string;
                }[];
                /**
                 * @description An abbreviated version of the error or warning in textual format.
                 * @example WARN.SP.PROVIDER_WARNING
                 */
                ShortText?: string;
                /**
                 * @description If present, this attribute may identify an unknown or misspelled tag that caused an error in processing.
                 *
                 *     It is recommended that the Tag attribute use XPath notation to identify the location of a tag in the event that more than one tag of the same name is present in the document.
                 *
                 *     Alternatively, the tag name alone can be used to identify missing data [Type=ReqFieldMissing].
                 * @example Unknown Element
                 */
                Element?: string;
                /**
                 * @description If present, this attribute allows for batch processing and the identification of the record that failed amongst a group of records.
                 *
                 *     This value may contain a concatenation of a unique failed transaction ID with specific record(s) associated with that transaction.
                 * @example Unknown RecordID
                 */
                RecordID?: string;
                /**
                 * Format: uri
                 * @description If present, this attribute refers to an online description of the error that occurred.
                 * @example https://developer.sabre.com
                 */
                DocURL?: string;
                /**
                 * Format: iso8601-date-time
                 * @description Transaction time stamp in `yyyy-mm-ddThh:mm:ss.SSSZ` format.
                 * @example 2026-02-13T04:16:10.174-06:00
                 */
                timeStamp?: string;
                /**
                 * @description If present, this attribute provides an XML IDREF to the element for which the results apply.
                 * @example XML IDREF
                 */
                reference?: string;
            }[];
            /**
             * @description An indication of the source of error or warning when processing the request.
             * @example BusinessLogic
             * @enum {string}
             */
            type?: "Transport" | "Validation" | "Application" | "BusinessLogic";
            /**
             * Format: iso8601-date-time
             * @description Transaction time stamp in "yyyy-mm-ddThh:mm:ss.SSSZ" format.
             * @example 2026-02-13T04:16:10.174-06:00
             */
            timeStamp?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}